+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -1,5 +1,6 @@
 package cart.controller;
 
+import cart.controller.api.ProductController;
 import cart.controller.dto.ProductRequest;
 import cart.service.ProductService;
 import com.fasterxml.jackson.databind.ObjectMapper;
