+++ src/test/java/cart/controller/ViewControllerTest.java
@@ -2,6 +2,7 @@ package cart.controller;
 
 import cart.controller.dto.MemberResponse;
 import cart.controller.dto.ProductResponse;
+import cart.controller.view.ViewController;
 import cart.domain.Member;
 import cart.domain.Product;
 import cart.service.MemberService;
